Added runtime tests for list, number and variant

+++ tests/runtime/variants/wasm.go
@@ -15,21 +15,21 @@ type VariantsImpl struct {
 
 func (i VariantsImpl) TestImports() {
 	{
-		res := ImportsRoundtripOption(Some[uint32](1))
+		res := ImportsRoundtripOption(Some[float32](1))
 		if res.IsNone() {
 			panic("ImportsRoundtripOption")
 		}
-		if res.Unwrap() != 1 {
+		if res.Unwrap() != 1.0 {
 			panic("ImportsRoundtripOption")
 		}
 
-		res2 := ImportsRoundtripOption(None[uint32]())
+		res2 := ImportsRoundtripOption(None[float32]())
 		if res2.IsSome() {
 			panic("ImportsRoundtripOption")
 		}
 	}
 	{
-		var param Result[uint32, uint32]
+		var param Result[uint32, float32]
 		param.Set(5)
 		res := ImportsRoundtripResult(param)
 		if res.IsErr() {
@@ -39,12 +39,12 @@ func (i VariantsImpl) TestImports() {
 			panic("ImportsRoundtripResult")
 		}
 
-		param.SetErr(10)
+		param.SetErr(10.0)
 		res2 := ImportsRoundtripResult(param)
 		if res2.IsOk() {
 			panic("ImportsRoundtripResult")
 		}
-		if res2.UnwrapErr() != 10 {
+		if res2.UnwrapErr() != 10.0 {
 			panic("ImportsRoundtripResult")
 		}
 	}
@@ -77,7 +77,7 @@ func (i VariantsImpl) TestImports() {
 		a.F2 = ImportsC3A(3)
 		a.F3 = ImportsC4A(4)
 		a.F4 = ImportsC5A(5)
-		a.F5 = ImportsC6A(6)
+		a.F5 = ImportsC6A(6.0)
 		res := ImportsVariantCasts(a)
 		if res.F0.GetA() != 1 {
 			panic("ImportsVariantCasts")
@@ -94,35 +94,35 @@ func (i VariantsImpl) TestImports() {
 		if res.F4.GetA() != 5 {
 			panic("ImportsVariantCasts")
 		}
-		if res.F5.GetA() != 6 {
+		if res.F5.GetA() != 6.0 {
 			panic("ImportsVariantCasts")
 		}
 	}
 	{
 		var a ImportsCasts
 		a.F0 = ImportsC1B(1)
-		a.F1 = ImportsC2B(2)
-		a.F2 = ImportsC3B(3)
-		a.F3 = ImportsC4B(4)
-		a.F4 = ImportsC5B(5)
-		a.F5 = ImportsC6B(6)
+		a.F1 = ImportsC2B(2.0)
+		a.F2 = ImportsC3B(3.0)
+		a.F3 = ImportsC4B(4.0)
+		a.F4 = ImportsC5B(5.0)
+		a.F5 = ImportsC6B(6.0)
 		res := ImportsVariantCasts(a)
 		if res.F0.GetB() != 1 {
 			panic("ImportsVariantCasts")
 		}
-		if res.F1.GetB() != 2 {
+		if res.F1.GetB() != 2.0 {
 			panic("ImportsVariantCasts")
 		}
-		if res.F2.GetB() != 3 {
+		if res.F2.GetB() != 3.0 {
 			panic("ImportsVariantCasts")
 		}
-		if res.F3.GetB() != 4 {
+		if res.F3.GetB() != 4.0 {
 			panic("ImportsVariantCasts")
 		}
-		if res.F4.GetB() != 5 {
+		if res.F4.GetB() != 5.0 {
 			panic("ImportsVariantCasts")
 		}
-		if res.F5.GetB() != 6 {
+		if res.F5.GetB() != 6.0 {
 			panic("ImportsVariantCasts")
 		}
 	}
@@ -130,8 +130,8 @@ func (i VariantsImpl) TestImports() {
 		var a ImportsZeros
 		a.F0 = ImportsZ1A(1)
 		a.F1 = ImportsZ2A(2)
-		a.F2 = ImportsZ3A(3)
-		a.F3 = ImportsZ4A(4)
+		a.F2 = ImportsZ3A(3.0)
+		a.F3 = ImportsZ4A(4.0)
 		res := ImportsVariantZeros(a)
 		if res.F0.GetA() != 1 {
 			panic("ImportsVariantZeros")
@@ -139,10 +139,10 @@ func (i VariantsImpl) TestImports() {
 		if res.F1.GetA() != 2 {
 			panic("ImportsVariantZeros")
 		}
-		if res.F2.GetA() != 3 {
+		if res.F2.GetA() != 3.0 {
 			panic("ImportsVariantZeros")
 		}
-		if res.F3.GetA() != 4 {
+		if res.F3.GetA() != 4.0 {
 			panic("ImportsVariantZeros")
 		}
 	}
@@ -188,7 +188,7 @@ func (i VariantsImpl) TestImports() {
 
 }
 
-func (i VariantsImpl) RoundtripOption(a Option[uint32]) Option[uint8] {
+func (i VariantsImpl) RoundtripOption(a Option[float32]) Option[uint8] {
 	if a.IsNone() {
 		return None[uint8]()
 	} else {
@@ -196,10 +196,10 @@ func (i VariantsImpl) RoundtripOption(a Option[uint32]) Option[uint8] {
 	}
 }
 
-func (i VariantsImpl) RoundtripResult(a Result[uint32, uint32]) Result[uint64, uint8] {
-	var res Result[uint64, uint8]
+func (i VariantsImpl) RoundtripResult(a Result[uint32, float32]) Result[float64, uint8] {
+	var res Result[float64, uint8]
 	if a.IsOk() {
-		res.Set(uint64(a.Unwrap()))
+		res.Set(float64(a.Unwrap()))
 	} else {
 		res.SetErr(uint8(a.UnwrapErr()))
 	}
